Move demo to Overview doc
